# api/openapi.json
{
  "components": {
    "schemas": {
...
      "ErrorResponse": {
        "additionalProperties": false,
        "description": "Standard envelope for all 4xx / 5xx responses.",
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorDetail"
          }
        },
        "required": [
          "error"
        ],
        "title": "ErrorResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "OwnershipRequestAdminCreate": {
        "additionalProperties": false,
        "description": "Admin-supplied body for creating an ownership request on someone's behalf.\n\nDifferent from the public ``OwnershipRequestCreate`` in two ways:\n  1. ``place_id`` is in the body instead of the URL \u2014 the admin UI\n     picks the place via a search widget rather than navigating to\n     the place detail page first.\n  2. ``requester_user_id`` is optional and admin-supplied. The\n     public endpoint derives this from the caller's auth; an admin\n     creating on behalf of someone else shouldn't end up as the\n     requester on the record. Leave null for unauthenticated\n     requesters (phone-call, in-person intake, etc.).",
        "properties": {
          "contact_email": {
            "format": "email",
            "maxLength": 255,
            "title": "Contact Email",
            "type": "string"
          },
          "contact_name": {
            "maxLength": 255,
            "minLength": 1,
            "title": "Contact Name",
            "type": "string"
          },
          "contact_phone": {
            "anyOf": [
              {
                "maxLength": 50,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Contact Phone"
          },
          "message": {
            "anyOf": [
              {
                "maxLength": 2000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message"
          },
          "place_id": {
            "format": "uuid",
            "title": "Place Id",
            "type": "string"
          },
          "requester_user_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Requester User Id"
          }
        },
        "required": [
          "place_id",
          "contact_name",
          "contact_email"
        ],
        "title": "OwnershipRequestAdminCreate",
        "type": "object"
      },
      "OwnershipRequestAdminRead": {
        "properties": {
          "contact_email": {
            "format": "email",
            "title": "Contact Email",
            "type": "string"
          },
          "contact_name": {
            "title": "Contact Name",
            "type": "string"
          },
          "contact_phone": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Contact Phone"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message"
          },
          "place_id": {
            "format": "uuid",
            "title": "Place Id",
            "type": "string"
          },
          "requester_user_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Requester User Id"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "id",
          "place_id",
          "requester_user_id",
          "contact_name",
          "contact_email",
          "contact_phone",
          "message",
          "status",
          "created_at",
          "updated_at"
        ],
        "title": "OwnershipRequestAdminRead",
        "type": "object"
      },
      "OwnershipRequestApprove": {
        "description": "Approve an ownership request.\n\nProvide either an existing `organization_id` OR `new_organization_name`\n(to create a new org on the fly). Exactly one is required.",
        "properties": {
          "member_role": {
            "default": "OWNER_ADMIN",
            "maxLength": 50,
            "title": "Member Role",
            "type": "string"
          },
          "new_organization_name": {
            "anyOf": [
              {
                "maxLength": 255,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "New Organization Name"
          },
          "note": {
            "anyOf": [
              {
                "maxLength": 2000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          },
          "organization_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Organization Id"
          },
          "place_owner_role": {
            "default": "PRIMARY",
            "maxLength": 50,
            "title": "Place Owner Role",
            "type": "string"
          }
        },
        "title": "OwnershipRequestApprove",
        "type": "object"
      },
...
      "OwnershipRequestEvidence": {
        "properties": {
          "note": {
            "anyOf": [
              {
                "maxLength": 2000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          }
        },
        "title": "OwnershipRequestEvidence",
        "type": "object"
      },
...
      "OwnershipRequestReject": {
        "properties": {
          "reason": {
            "maxLength": 2000,
            "minLength": 3,
            "title": "Reason",
            "type": "string"
          }
        },
        "required": [
          "reason"
        ],
        "title": "OwnershipRequestReject",
        "type": "object"
      },
...
      "PlaceOwnerAdminRead": {
        "description": "One ``place_owners`` link row, admin-facing.\n\nReturned from ``GET /admin/places/{id}/owners``. The admin UI renders\none of these per row on the place detail page, ordered ACTIVE first\nso the \"who's actually managing this place today\" answer is at the top.",
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "organization": {
            "$ref": "#/components/schemas/OrganizationSummaryAdmin"
          },
          "role": {
            "title": "Role",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "id",
          "organization",
          "role",
          "status",
          "created_at"
        ],
        "title": "PlaceOwnerAdminRead",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "TrustHalal API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/admin/ownership-requests": {
      "get": {
        "operationId": "list_ownership_requests_admin_ownership_requests_get",
        "parameters": [
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 50,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "exclusiveMinimum": 0,
              "maximum": 200,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "X-User-Id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-User-Id"
            }
          },
          {
            "in": "cookie",
            "name": "tht_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Hht Session"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/OwnershipRequestAdminRead"
                  },
                  "title": "Response List Ownership Requests Admin Ownership Requests Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Ownership Requests",
        "tags": [
          "admin"
        ]
      },
      "post": {
        "description": "Create an ownership request on someone's behalf (admin path).\n\nUse when an admin takes an inbound request by phone, email, or\nin-person and wants to capture it in the system without the\nclaimant going through the public submit flow. ``requester_user_id``\ncan be null for unauthenticated intakes; set it to a real user id\nwhen the claimant has an account and you want them to be able to\nsee the request later via ``GET /ownership-requests/{id}/detail``.",
        "operationId": "create_ownership_request_admin_admin_ownership_requests_post",
        "parameters": [
          {
            "in": "header",
            "name": "X-User-Id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-User-Id"
            }
          },
          {
            "in": "cookie",
            "name": "tht_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Hht Session"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OwnershipRequestAdminCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OwnershipRequestAdminRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Ownership Request Admin",
        "tags": [
          "admin"
        ]
      }
    },
    "/admin/ownership-requests/{request_id}/approve": {
      "post": {
        "operationId": "approve_ownership_request_admin_ownership_requests__request_id__approve_post",
        "parameters": [
          {
            "in": "path",
            "name": "request_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Request Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-User-Id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-User-Id"
            }
          },
          {
            "in": "cookie",
            "name": "tht_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Hht Session"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OwnershipRequestApprove"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OwnershipRequestAdminRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Approve Ownership Request",
        "tags": [
          "admin"
        ]
      }
    },
    "/admin/ownership-requests/{request_id}/reject": {
      "post": {
        "operationId": "reject_ownership_request_admin_ownership_requests__request_id__reject_post",
        "parameters": [
          {
            "in": "path",
            "name": "request_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Request Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-User-Id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-User-Id"
            }
          },
          {
            "in": "cookie",
            "name": "tht_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Hht Session"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OwnershipRequestReject"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OwnershipRequestAdminRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Reject Ownership Request",
        "tags": [
          "admin"
        ]
      }
    },
    "/admin/ownership-requests/{request_id}/request-evidence": {
      "post": {
        "operationId": "request_more_evidence_admin_ownership_requests__request_id__request_evidence_post",
        "parameters": [
          {
            "in": "path",
            "name": "request_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Request Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-User-Id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-User-Id"
            }
          },
          {
            "in": "cookie",
            "name": "tht_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Hht Session"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OwnershipRequestEvidence"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OwnershipRequestAdminRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Request More Evidence",
        "tags": [
          "admin"
        ]
      }
    },
...
    "/admin/places/{place_id}/owners": {
      "get": {
        "description": "List organization ownership links for a place.\n\nReturns every ``place_owners`` row for the given place, joined with\nthe owning organization and the org's active-member count. Sorted\nACTIVE-first so the admin UI shows \"who's running this today\" before\nhistorical or pending links.\n\nEmpty list is a valid response \u2014 the place simply has no owners yet.\nSoft-deleted places still resolve here; the admin needs to see\nownership context when triaging restore decisions.",
        "operationId": "list_place_owners_admin_admin_places__place_id__owners_get",
        "parameters": [
          {
            "in": "path",
            "name": "place_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Place Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-User-Id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-User-Id"
            }
          },
          {
            "in": "cookie",
            "name": "tht_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Hht Session"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/PlaceOwnerAdminRead"
                  },
                  "title": "Response List Place Owners Admin Admin Places  Place Id  Owners Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Missing or invalid X-User-Id header."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Caller is authenticated but not an admin."
          },
          "404": {
            "content": {
              "application/json": {
                "example": {
                  "error": {
                    "code": "PLACE_NOT_FOUND",
                    "message": "Place not found"
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "No place with this id exists (even in the soft-deleted set)."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Place Owners Admin",
        "tags": [
          "admin"
        ]
      }
    },
    "/admin/places/{place_id}/owners/{owner_id}": {
      "delete": {
        "description": "Revoke a place's ownership relationship.\n\nSoft-unlink: flips the ``PlaceOwner.status`` to ``REVOKED`` rather\nthan deleting the row. The schema's partial unique index on\n``place_id WHERE status IN ('PENDING','ACTIVE','VERIFIED')``\nexcludes REVOKED, so the place becomes eligible for a fresh live\nowner after this action \u2014 and the historical row survives so\n\"Acme Inc used to own this place\" remains answerable from the\nevent history.\n\nOptional ``reason`` body flows to the EDITED PlaceEvent message\nalongside the org name + role + prior status.\n\nIdempotent: revoking an already-REVOKED owner is a 204 no-op (no\nnew event logged).",
        "operationId": "revoke_place_owner_admin_admin_places__place_id__owners__owner_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "place_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Place Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "owner_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Owner Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-User-Id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-User-Id"
            }
          },
          {
            "in": "cookie",
            "name": "tht_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Hht Session"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "$ref": "#/components/schemas/PlaceOwnerRevokeRequest"
                  },
                  {
                    "type": "null"
                  }
                ],
                "title": "Payload"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Missing/invalid X-User-Id."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not an admin."
          },
          "404": {
            "content": {
              "application/json": {
                "example": {
                  "error": {
                    "code": "OWNERSHIP_NOT_FOUND",
                    "message": "No ownership link with this id on this place"
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Either the place id is unknown (PLACE_NOT_FOUND) or the owner_id isn't a PlaceOwner row for this place (OWNERSHIP_NOT_FOUND). Requiring both guards against typoing the place_id in the URL and accidentally revoking an ownership row on a different place."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Reason failed validation (too short or too long)."
          }
        },
        "summary": "Revoke Place Owner Admin",
        "tags": [
          "admin"
        ]
      }
    },
...
    "/ownership-requests/{request_id}": {
      "get": {
        "operationId": "get_ownership_request_status_ownership_requests__request_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "request_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Request Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OwnershipRequestStatusRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Ownership Request Status",
        "tags": [
          "ownership-requests"
        ]
      }
    },
...
    "/places/{place_id}/ownership-requests": {
      "post": {
        "operationId": "submit_ownership_request_places__place_id__ownership_requests_post",
        "parameters": [
          {
            "in": "path",
            "name": "place_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Place Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-User-Id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-User-Id"
            }
          },
          {
            "in": "cookie",
            "name": "tht_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Hht Session"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OwnershipRequestCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OwnershipRequestRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Submit Ownership Request",
        "tags": [
          "ownership-requests"
        ]
      }
    }
  }
}